Zadanie1 rekrutacyjne - Ładowanie wszystkich commitów ze wszystkich gałęzi. poprawki CD ... 5 i poprawki widoczności przycisków.

+++ Soneta.Examples/Zadanie1/Extender/PolaListyComitow.cs
@@ -14,7 +14,7 @@ namespace Soneta.Examples.Zadanie1.Extender
         public string Autor { get; set; }
         public DateTime Data { get; set; }
         public string Opis { get; set; }
-        public int Ilosc { get; set; }
+        public decimal Ilosc { get; set; }
     }
     public class PolaListyBranches
     {
